rename gmpb plugin to modes
