--- gdsfactory/simulation/gmpb/__init__.py
@@ -1,27 +1,0 @@
-from gdsfactory.simulation.gmpb import coupler, waveguide
-from gdsfactory.simulation.gmpb.find_coupling_vs_gap import (
-    find_coupling_vs_gap,
-    plot_coupling_vs_gap,
-)
-from gdsfactory.simulation.gmpb.find_mode_dispersion import find_mode_dispersion
-from gdsfactory.simulation.gmpb.find_modes import find_modes
-from gdsfactory.simulation.gmpb.find_neff_vs_width import (
-    find_neff_vs_width,
-    plot_neff_vs_width,
-)
-from gdsfactory.simulation.gmpb.get_mode_solver_coupler import get_mode_solver_coupler
-from gdsfactory.simulation.gmpb.get_mode_solver_rib import get_mode_solver_rib
-
-__all__ = [
-    "find_modes",
-    "find_neff_vs_width",
-    "find_mode_dispersion",
-    "find_coupling_vs_gap",
-    "get_mode_solver_rib",
-    "get_mode_solver_coupler",
-    "plot_neff_vs_width",
-    "plot_coupling_vs_gap",
-    "coupler",
-    "waveguide",
-]
-__version__ = "0.0.2"
