--- gdsfactory/simulation/gmpb/test_dw_dh/test_dw_dh.csv
@@ -1,10 +1,0 @@
-,dw,dh,neff,ng
-0,-0.029999999999999999,-0.02,2.2079814942226128,4.2082883265980282
-1,-0.029999999999999999,0,2.2911352776810316,4.2630678678062619
-2,-0.029999999999999999,0.02,2.3706531151244499,4.3068679645291752
-3,0,-0.02,2.2776910458764195,4.172520415440367
-4,0,0,2.3606235599852794,4.2163468017815005
-5,0,0.02,2.4397680152547889,4.2523488071312858
-6,0.029999999999999999,-0.02,2.3275781528651227,4.1224942052289411
-7,0.029999999999999999,0,2.4099534941389762,4.1612808707676674
-8,0.029999999999999999,0.02,2.4883754289663194,4.1941979995121494
