--- gdsfactory/simulation/gmpb/test_dw_dh/test_dw_dh_dispersion.csv
@@ -1,10 +1,0 @@
-,dw,dh,neff,ng
-0,-0.029999999999999999,-0.02,2.206453004523361,4.2079913957808488
-1,-0.029999999999999999,0,2.3023352525535103,4.2910429467043087
-2,-0.029999999999999999,0.02,2.3679225676233777,4.3040893844495329
-3,0,-0.02,2.2765778913596897,4.1696638779837958
-4,0,0,2.3726404761456306,4.2395646987016189
-5,0,0.02,2.4369416740622127,4.2479561297978323
-6,0.029999999999999999,-0.02,2.3266718962571367,4.1216226181828475
-7,0.029999999999999999,0,2.4219271472867696,4.1850232220839985
-8,0.029999999999999999,0.02,2.4860439692855025,4.1911829721287823
