+++ gdsfactory/simulation/modes/find_neff_ng_dw_dh.py
@@ -14,7 +14,7 @@ import pydantic
 from scipy.interpolate import interp2d
 
 from gdsfactory.config import PATH
-from gdsfactory.simulation.gmpb.find_mode_dispersion import find_mode_dispersion
+from gdsfactory.simulation.modes.find_mode_dispersion import find_mode_dispersion
 
 PATH.modes = pathlib.Path.cwd() / "data"
