+++ gdsfactory/simulation/modes/test_dw_dh.py
@@ -1,4 +1,4 @@
-from gdsfactory.simulation.gmpb.find_neff_ng_dw_dh import find_neff_ng_dw_dh
+from gdsfactory.simulation.modes.find_neff_ng_dw_dh import find_neff_ng_dw_dh
 
 
 def test_dw_dh(dataframe_regression):
