+++ gdsfactory/simulation/modes/types.py
@@ -203,7 +203,7 @@ ModeSolverOrFactory = Union[mpb.ModeSolver, ModeSolverFactory]
 
 
 if __name__ == "__main__":
-    # import gdsfactory.simulation.gmpb as gm
+    # import gdsfactory.simulation.modes as gm
     # m = gm.find_modes()
     # m[1].plot_e_all()
     w = Waveguide()
